Modified naming conventions
