+++ src/main/java/com/nha/abdm/wrapper/hip/hrp/dataTransfer/DataTransferService.java
@@ -2,35 +2,50 @@
 package com.nha.abdm.wrapper.hip.hrp.dataTransfer;
 
 import com.nha.abdm.wrapper.common.responses.GatewayCallbackResponse;
-import com.nha.abdm.wrapper.hip.hrp.dataTransfer.responses.BundleResponseHIP;
-import com.nha.abdm.wrapper.hip.hrp.dataTransfer.responses.DataNotifyResponse;
-import com.nha.abdm.wrapper.hip.hrp.dataTransfer.responses.DataRequestResponse;
+import com.nha.abdm.wrapper.hip.hrp.dataTransfer.requests.callback.BundleResponseHIP;
+import com.nha.abdm.wrapper.hip.hrp.dataTransfer.requests.callback.HIPConsentNotification;
+import com.nha.abdm.wrapper.hip.hrp.dataTransfer.requests.callback.HIPHealthInformationRequest;
 import org.springframework.stereotype.Service;
 
 @Service
 public class DataTransferService implements DataTransferInterface {
+  /**
+   * The callback from ABDM gateway after consentGrant for dataTransfer,
+   * POST method for /on-notify
+   * @param HIPConsentNotification careContext and demographics details are provided, and implement a logic to check the existence of the careContexts.
+   */
   @Override
-  public GatewayCallbackResponse DataOnNotifyCall(DataNotifyResponse dataNotifyResponse) {
+  public void notifyOnReceived(HIPConsentNotification HIPConsentNotification) {
     // TODO Implement onNotify workflow logic.
-    return null;
   }
 
+  /**
+   *
+   * POST /on-request as an acknowledgement for agreeing to make dataTransfer to ABDM gateway.
+   * @param HIPHealthInformationRequest HIU public keys and dataPush URL is provided
+   */
   @Override
-  public GatewayCallbackResponse DataOnRequestCall(DataRequestResponse dataRequestResponse) {
+  public void requestOnReceived(HIPHealthInformationRequest HIPHealthInformationRequest) {
     // TODO Implement onRequest workflow logic.
-    return null;
+    initiateBundleRequest(HIPHealthInformationRequest);
   }
 
+  /**
+   * Requesting HIP for FHIR bundle
+   * @param HIPHealthInformationRequest use the requestId to fetch the careContexts from dump to request HIP.
+   */
   @Override
-  public GatewayCallbackResponse initiateBundleRequest(DataRequestResponse dataRequestResponse) {
+  public void initiateBundleRequest(HIPHealthInformationRequest HIPHealthInformationRequest) {
     // TODO Implement bundle request to HIP workflow logic.
-    return null;
   }
 
+  /**
+   * Encrypt the bundle and
+   * POST to /dataPushUrl of HIU
+   * @param bundleResponseHIP FHIR bundle received from HIP for the particular patients
+   */
   @Override
-  public GatewayCallbackResponse initiateDataTransfer(BundleResponseHIP bundleResponseHIP)
-      throws Exception {
+  public void initiateDataTransfer(BundleResponseHIP bundleResponseHIP) {
     // TODO Implement dataPush workflow logic.
-    return null;
   }
 }
