+++ src/main/java/com/nha/abdm/wrapper/hip/hrp/dataTransfer/requests/HIPConsentNotificationResponse.java
@@ -1,0 +1,22 @@
+/* (C) 2024 */
+package com.nha.abdm.wrapper.hip.hrp.dataTransfer.requests;
+
+import com.nha.abdm.wrapper.common.models.RespRequest;
+import com.nha.abdm.wrapper.common.responses.ErrorResponse;
+import com.nha.abdm.wrapper.hip.hrp.dataTransfer.requests.helpers.ConsentAcknowledgement;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class HIPConsentNotificationResponse {
+  public String requestId;
+  public String timestamp;
+  public ConsentAcknowledgement acknowledgement;
+  public ErrorResponse error;
+  public RespRequest resp;
+}
