+++ src/main/java/com/nha/abdm/wrapper/hip/hrp/dataTransfer/requests/callback/BundleResponseHIP.java
@@ -1,0 +1,16 @@
+/* (C) 2024 */
+package com.nha.abdm.wrapper.hip.hrp.dataTransfer.requests.callback;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Data
+@Builder
+public class BundleResponseHIP {
+  String careContextReference;
+  String bundleContent;
+}
