+++ src/main/java/com/nha/abdm/wrapper/hip/hrp/dataTransfer/requests/callback/helpers/DataHiRequest/DatakeyMaterial/DataDhPublicKey.java
@@ -1,0 +1,18 @@
+/* (C) 2024 */
+package com.nha.abdm.wrapper.hip.hrp.dataTransfer.requests.callback.helpers.DataHiRequest.DatakeyMaterial;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Data
+@Builder
+public class DataDhPublicKey {
+
+  public String expiry;
+  public String keyValue;
+  public String parameters;
+}
