+++ src/main/java/com/nha/abdm/wrapper/hip/hrp/dataTransfer/requests/callback/helpers/DataNotification/DataConsentdetail/DataPatient/DataPatient.java
@@ -1,0 +1,15 @@
+/* (C) 2024 */
+package com.nha.abdm.wrapper.hip.hrp.dataTransfer.requests.callback.helpers.DataNotification.DataConsentdetail.DataPatient;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Data
+@Builder
+public class DataPatient {
+  public String id;
+}
