--- src/main/java/com/nha/abdm/wrapper/hip/hrp/dataTransfer/requests/helpers/DataHiStatus.java
@@ -1,16 +1,0 @@
-/* (C) 2024 */
-package com.nha.abdm.wrapper.hip.hrp.dataTransfer.requests.helpers;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class DataHiStatus {
-  public String transactionId;
-  public String sessionStatus;
-}
