--- src/main/java/com/nha/abdm/wrapper/hip/hrp/dataTransfer/responses/helpers/DataNotification/DataConsentdetail/DataHip/DataHip.java
@@ -1,16 +1,0 @@
-/* (C) 2024 */
-package com.nha.abdm.wrapper.hip.hrp.dataTransfer.responses.helpers.DataNotification.DataConsentdetail.DataHip;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Data
-@Builder
-public class DataHip {
-  public String id;
-  public String name;
-}
