--- src/main/java/com/nha/abdm/wrapper/hip/hrp/dataTransfer/responses/helpers/DataNotification/DataConsentdetail/DataPermission/DataDateRange.java
@@ -1,16 +1,0 @@
-/* (C) 2024 */
-package com.nha.abdm.wrapper.hip.hrp.dataTransfer.responses.helpers.DataNotification.DataConsentdetail.DataPermission;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Data
-@Builder
-public class DataDateRange {
-  public String from;
-  public String to;
-}
